requirements.txt
aniso8601==7.0.0
APScheduler==3.6.1
asn1crypto==0.24.0
autopep8==1.4.4
cffi==1.12.3
cryptography==2.7
Django==2.2.4
django-apscheduler==0.3.0
django-cors-headers==3.1.0
django-filter==2.2.0
django-markdown-deux==1.0.5
django-rest-knox==4.1.0
djangorestframework==3.10.2
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
markdown2==2.3.8
promise==2.2.1
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
tzlocal==2.0.0
whitenoise==5.0.1
django-admin-rangefilter
